Only allow single thread for find_essential_reactions()
A warning message has been added for when multiple threads are asked
for. Due to the set up of PyFBA, the GLPK solver is set as a global
variable that may be compromised during multi-threading -- a thread
could edit the solver matrix, reaction bounds, and/or compound bounds
while it is in use of another thread. A refactoring of the code is
necessary to enable concurrent solvers, probably one for each FBA
instance.